requirements.txt
python-Wappalyzer
argparse
beautifulsoup4
mechanize
requests
progress
dnspython
shodan
nmap